[FIX] account, orm, website_sale, cloud_storage_migration: stop narrowing ids
Record ids are 64-bit on databases created with bigint identifiers, and these
places still handled them as 32-bit integers.

`account.move._compute_payment_state` called `BOOL(pay.id)` to test whether a
payment exists. PostgreSQL defines a cast from int4 to bool but none from
int8, so the query failed outright with "function bool(bigint) does not
exist". Testing the column against NULL says what the code means and works
whatever the width is.

The remaining ones cast ids down to a 32-bit integer, which raises "integer
out of range" once an id exceeds 2^31:

- account.move._auto_post_draft_entries builds its VALUES list with ::int4
- the ORM casts the jsonb values of a company-dependent many2one with ::int4
  when unlinking the records they point to
- website_sale splits product_public_category.parent_path into ::integer
- cloud_storage_migration reads an attachment id from ir.config_parameter

+++ odoo/orm/models.py
@@ -3661,8 +3661,8 @@ class BaseModel(metaclass=MetaModel):
                             SELECT jsonb_object_agg(
                                 key,
                                 CASE
-                                    WHEN value::int4 in %(ids)s THEN NULL
-                                    ELSE value::int4
+                                    WHEN value::int8 in %(ids)s THEN NULL
+                                    ELSE value::int8
                                 END)
                             FROM jsonb_each_text(%(field)s)
                         )
